requirements.txt
asgiref==3.5.2
autopep8==1.7.0
beautifulsoup4==4.11.1
Django==3.2.15
django-bootstrap4==22.2
django-crispy-forms==1.14.0
django-forms-bootstrap==3.1.0
importlib-metadata==2.1.3
Pillow==9.2.0
pycodestyle==2.9.1
pytz==2022.4
soupsieve==2.3.2.post1
sqlparse==0.4.3
toml==0.10.2
typing-extensions==4.3.0
zipp==3.8.1

django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==2.1.0
